Stop putting directories from the file names onto the import path
This behavior was incorrect and led to strange behaviors. Proto files are meant
to be importable using hierarchical names which should be found relative to the
import paths. Hence, adding the directory part of the given path names to the
import path only encourages incorrect use of hierarchical names

For example, if my proto file is named api.proto and is not meant to be
imported with a hierarchical name, it is incorrect to allow me to specify it as
`path/to/api.proto` because if I import a second file which happens to import
`api.proto` the parser has no way of knowing that it is supposed to be the same
as `path/to/api.proto` which I originally gave it.

Addresses issue #80